% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asympVariance}
\alias{asympVariance}
\title{Compute asymptotic variance of a moment sequence (weights, support)}
\usage{
asympVariance(weights, support)
}
\arguments{
\item{weights}{weights of a moment sequence}

\item{support}{support of a moment sequence}
}
\description{
Compute asymptotic variance of a moment sequence (weights, support)
}
