% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avar}
\alias{avar}
\title{Compute asymptotic variance of a moment sequence}
\usage{
avar(SRfit1)
}
\arguments{
\item{SRfit}{momentLS fit object}
}
\description{
Compute asymptotic variance of a moment sequence
}
