% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeMoments}
\alias{computeMoments}
\title{compute first M moments of a moment sequence (weights, support)}
\usage{
computeMoments(support, weights, M = 100)
}
\description{
compute first M moments of a moment sequence (weights, support)
}
