% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{computeXtr_w}
\alias{computeXtr_w}
\title{compute XtX_w}
\usage{
computeXtr_w(
  alphaGrid,
  r,
  m_uw,
  numchck_control = list(chck_TOL = .Machine$double.eps^{
     1/2
 }, rplc_TOL =
    .Machine$double.eps^{
     1/3
 }, N_wseq = 1e+06)
)
}
\description{
compute XtX_w
}
