% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{innerProduct_L2Moment}
\alias{innerProduct_L2Moment}
\title{inner product between r in L2(Z) and m in M_infty}
\usage{
innerProduct_L2Moment(
  r,
  support,
  weights,
  precomputed = list(norm2_r = NULL, X_unscl_Atr = NULL)
)
}
\description{
inner product between r in L2(Z) and m in M_infty
}
