% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{innerProduct_twoMoments}
\alias{innerProduct_twoMoments}
\title{inner product between m1,m2 in M_infty}
\usage{
innerProduct_twoMoments(support1, weights1, support2, weights2)
}
\description{
inner product between m1,m2 in M_infty
}
