% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{L2diff_L2Moment}
\alias{L2diff_L2Moment}
\title{compute || r - m ||^2 for r in L2(Z) and m in M_infty}
\usage{
L2diff_L2Moment(
  r,
  support,
  weights,
  precomputed = list(norm2_r = NULL, X_unscl_Atr = NULL)
)
}
\description{
compute || r - m ||^2 for r in L2(Z) and m in M_infty
}
