% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{L2diff_twoMoments}
\alias{L2diff_twoMoments}
\title{compute || m1 - m2 ||^2 for m1,m2 in M_infty}
\usage{
L2diff_twoMoments(support1, weights1, support2, weights2)
}
\description{
compute || m1 - m2 ||^2 for m1,m2 in M_infty
}
