% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{makeXtX_w}
\alias{makeXtX_w}
\title{compute Xtr_w}
\usage{
makeXtX_w(
  alphaGrid,
  m_uw,
  numchck_control = list(chck_TOL = .Machine$double.eps^{
     1/2
 }, rplc_TOL =
    .Machine$double.eps^{
     1/3
 }, N_wseq = 1e+06)
)
}
\description{
compute Xtr_w
}
