\name{momentLS-package}
\alias{momentLS-package}
\alias{momentLS}
\docType{package}
\title{
\packageTitle{momentLS}
}
\description{
\packageDescription{momentLS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{momentLS}
\packageIndices{momentLS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{momentLS}

Maintainer: \packageMaintainer{momentLS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
