% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentLS.R
\docType{package}
\name{momentLS}
\alias{momentLS}
\title{momentLS}
\description{
Implementation of Moment LS estimators in "Efficient shape-constrained inference for the autocovariance sequence from a reversible Markov chain" <https://arxiv.org/abs/2207.12705>
}
