% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR1.R
\name{SR1}
\alias{SR1}
\title{SR1}
\usage{
SR1(
  r,
  delta = NULL,
  alphaGrid = NULL,
  init = NULL,
  tol = 10^-8,
  maxit = 1000,
  gradTrace = FALSE,
  n_alphas = 1001,
  gridControl = list(cm = FALSE, scale = "log"),
  precomputed = list(s_alpha = NULL, XtX = NULL, Xtr = NULL, alphaGrid = NULL, input =
    NULL)
)
}
\arguments{
\item{r}{input autocovariance estimate}

\item{delta}{parameter to specify appropriate moment space; use tune_delta if unknown}

\item{alphaGrid}{a grid of x values from [-1+delta,1-delta]; will be automatically created if left unspecified}

\item{init}{initialization for the SR algorithm}

\item{tol}{algorithm tolerance}

\item{maxit}{maximum number of iterations for support reduction algorithm}

\item{gradTrace}{trace gradients if TRUE}

\item{gridControl}{list which contains parameters to specify alphaGrid}

\item{precomputed}{an optional list which can contain some precomputed quantities for computational speed-up}
}
\value{
a SR1fit object containing estimated support and weights
}
\description{
Fit the MomentLS estimator given the input autocovariance sequence estimator
}
