% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asympVariance}
\alias{asympVariance}
\title{compute asymptotic variance of a moment sequence (weights, support)}
\usage{
asympVariance(weights, support)
}
\description{
compute asymptotic variance of a moment sequence (weights, support)
}
