% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocov.R
\name{autocov}
\alias{autocov}
\title{Compute the empirical autocovariances of chain x}
\usage{
autocov(x, center = TRUE)
}
\arguments{
\item{x}{input chain}

\item{center}{a logical value; TRUE for empirical mean centering}
}
\description{
Compute the empirical autocovariances of chain x
}
