% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeMoments}
\alias{computeMoments}
\title{Compute first M moments of a moment sequence (weights, support)}
\usage{
computeMoments(support, weights, M = 100)
}
\arguments{
\item{support}{support of a moment sequence}

\item{weights}{weights of a moment sequence}
}
\description{
Compute first M moments of a moment sequence (weights, support)
}
