% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_delta.R
\name{tune_delta}
\alias{tune_delta}
\title{Tune delta for momentLS estimators}
\usage{
tune_delta(r, method = "ft", seq_type = "average", const = 0.5, c1 = 1, c2 = 1)
}
\description{
Tune based on a modification of an adaptive bandwidth selection method by Politis <https://doi.org/10.1080/10485250310001604659>. See also our paper <https://arxiv.org/abs/2207.12705>.
}
